health-backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
psycopg2-binary==2.9.9
pydantic==2.9.2
bleak==0.22.3
cryptography==43.0.1
dbus-fast==2.24.1
